backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
email-validator==2.1.0.post1
SQLAlchemy==2.0.23
Werkzeug==2.3.7
alembic==1.12.1
gunicorn==21.2.0
geopy==2.4.1
redis==5.0.1
Flask-SocketIO==5.3.6 